Separated db internals (db, dao, entities) and api (repo, models) into different modules
