+++ core/db-api/src/main/java/com/mattrobertson/greek/reader/db/api/models/GlossModel.kt
@@ -1,0 +1,7 @@
+package com.mattrobertson.greek.reader.db.api.models
+
+data class GlossModel (
+	val lex: String,
+	val gloss: String,
+	val occ: Int
+)
