+++ core/db-api/src/main/java/com/mattrobertson/greek/reader/db/api/models/VerseModel.kt
@@ -1,0 +1,8 @@
+package com.mattrobertson.greek.reader.db.api.models
+
+data class VerseModel (
+	val book: Int,
+	val chapter: Int,
+	val verse: Int,
+	val encodedText: String
+)
