+++ core/db-internal/build.gradle.kts
@@ -1,0 +1,49 @@
+plugins {
+    id("com.android.library")
+    kotlin("android")
+    kotlin("kapt")
+}
+
+android {
+    compileSdk = Versions.compileSdk
+    buildToolsVersion = Versions.buildTools
+
+    defaultConfig {
+        minSdk = Versions.minSdk
+        targetSdk = Versions.targetSdk
+
+        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+    }
+
+    buildTypes {
+        getByName("release") {
+            isMinifyEnabled = false
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility = Versions.javaCompatibility
+        targetCompatibility = Versions.javaCompatibility
+    }
+
+    kotlinOptions {
+        jvmTarget = Versions.jvmTarget
+    }
+}
+
+dependencies {
+    implementation(project(":core:verseref"))
+
+    // Hilt
+    implementation("com.google.dagger:hilt-android:${Versions.hilt}")
+    kapt("com.google.dagger:hilt-android-compiler:${Versions.hilt}")
+
+    // Room
+    api("androidx.room:room-runtime:${Versions.room}")
+    api("androidx.room:room-ktx:${Versions.room}")
+    kapt("androidx.room:room-compiler:${Versions.room}")
+
+    // Testing
+    testImplementation("androidx.room:room-testing:${Versions.room}")
+    testImplementation("junit:junit:${Versions.junit}")
+}
