--- core/db/src/main/java/com/mattrobertson/greek/reader/db/repo/ConcordanceRepo.kt
@@ -1,9 +1,0 @@
-package com.mattrobertson.greek.reader.db.repo
-
-import com.mattrobertson.greek.reader.db.dao.ConcordanceDao
-
-class ConcordanceRepo(private val concordanceDao: ConcordanceDao) {
-
-	suspend fun getConcordanceEntries(lex: String) = concordanceDao.getConcordanceEntries(lex)
-
-}
