--- core/db/src/main/java/com/mattrobertson/greek/reader/db/repo/GlossesRepo.kt
@@ -1,9 +1,0 @@
-package com.mattrobertson.greek.reader.db.repo
-
-import com.mattrobertson.greek.reader.db.dao.GlossesDao
-
-class GlossesRepo(private val glossesDao: GlossesDao) {
-
-	suspend fun getGloss(lex: String) = glossesDao.getGloss(lex)
-
-}
