+++ feature/concordance/src/main/java/com/mattrobertson/greek/reader/concordance/ui/Concordance.kt
@@ -19,8 +19,8 @@ import androidx.compose.ui.text.style.TextDecoration
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.mattrobertson.greek.reader.db.repo.ConcordanceRepo
-import com.mattrobertson.greek.reader.db.repo.VerseRepo
+import com.mattrobertson.greek.reader.db.api.repo.ConcordanceRepo
+import com.mattrobertson.greek.reader.db.api.repo.VerseRepo
 import com.mattrobertson.greek.reader.ui.theme.SblGreek
 import com.mattrobertson.greek.reader.verseref.Book
 import com.mattrobertson.greek.reader.verseref.VerseRef
